add confirm dialog before message deletion

+++ client/src/actions/messagesActions.js
@@ -41,14 +41,16 @@ export function addMessage(message) {
 
 export function deleteMessage(messageId) {
     return (dispatch) => {
-        dispatch({ type: 'DELETING_MESSAGE' });
-        return fetch(`/messages/${messageId}`, {
-            method: 'DELETE'
-          }).then(response => {
-            // console.log(response);
-            if (response.ok) {
-                dispatch({ type: 'DELETE_MESSAGE_SUCCES', deletedMessageId: messageId })
-            }
-          })
+        if (window.confirm('Are you sure you want to delete this message?')) {
+            dispatch({ type: 'DELETING_MESSAGE' });
+            return fetch(`/messages/${messageId}`, {
+                method: 'DELETE'
+            }).then(response => {
+                // console.log(response);
+                if (response.ok) {
+                    dispatch({ type: 'DELETE_MESSAGE_SUCCES', deletedMessageId: messageId })
+                }
+            })
+        }
     };
 };
